+++ .pre-commit-config.yaml
@@ -1,0 +1,17 @@
+repos:
+  - repo: local
+    hooks:
+      - id: mypy
+        name: Validate types with MyPy
+        entry: poetry run mypy
+        language: system
+        types: [python]
+        pass_filenames: false
+        args:
+          - "benchmarktools"
+
+      - id: trunk
+        name: Format code with Trunk
+        entry: ./dev/trunk
+        language: system
+        args: ["fmt", "--no-progress"]
