+++ pyproject.toml
@@ -1,0 +1,24 @@
+[tool.poetry]
+name = "benchmark-tools"
+license="Apache License 2.0"
+version = "0.4.0"
+authors = ["BonvinLab <software.csb@gmail.com>"]
+description = "benchmarking framework for HADDOCK"
+packages = [
+    { include = "benchmarktools" },]
+
+[tool.poetry.dependencies]
+python = "^3.10"
+toml = "^0.10.2"
+
+[tool.poetry.dev-dependencies]
+mypy = "^0.971"
+pre-commit = "^2.20.0"
+pytest = "^7.1.3"
+
+[tool.poetry.scripts]
+haddock_bm = 'benchmarktools.cli:maincli'
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
